requirements.txt
coveralls==1.6.0
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-extensions==2.1.2
django-rest-swagger==2.2.0
djangorestframework==3.8.2
psycopg2==2.7.7
python-decouple==3.1
djangorestframework==3.8.2
flake8==3.7.7
PyJWT==1.7.1
pylint==2.2.2
requests==2.21.0
simplejson==3.16.0
uritemplate==3.0.0
django-heroku==0.3.1
gunicorn==19.9.0
heroku==0.1.4
Pillow==5.4.1
django-coverage-plugin==1.6.0
